# faster bin index for uniform edge