@Moelf 2020-10-29
notice this also makes multi-dimensional faster automatically. I hope @ianshmean is as happy as I am rn ;)

```
julia> d = rand(3,10^6);

julia> @btime fit(Histogram, (d[1,:], d[2,:], d[3,:]), (0:0.05:1, 0:0.05:1, 0:0.05:1));
  88.310 ms (15 allocations: 22.95 MiB)

julia> @btime fit(Histogram, (d[1,:], d[2,:], d[3,:]), (0:0.05:1, 0:0.05:1, 0:0.05:1));
  48.421 ms (15 allocations: 22.95 MiB)
```