@Octogonapus 2020-10-30
I gave this a try locally and it is a performance regression in our use case. Our bins are a `StepRange{UInt8,UInt8} 0x00:0x10:0xf0` so maybe something about the non-unit step causes this? I also had to clamp `n` between the bounds of the edge range to avoid an OOB exception (caused by the non-unit step in the range): `n = clamp(round(Int, (x - first(edge)) / step(edge) + 1), 1, length(edge))`.