@Octogonapus 2020-10-30
```julia
using Images, TestImages, FileIO, Rotations, CoordinateTransformations, StatsBase, BenchmarkTools

img = Gray.(testimage("lighthouse"));
tfm = Translation(125,250) ∘ LinearMap(RotMatrix(pi/50)) ∘ Translation(-125,-250);
img_rotated = warp(img, tfm);

fixed = collect(reinterpret(UInt8, img[50:300, 50:300]));
moving = collect(reinterpret(UInt8, img_rotated[50:300, 50:300]));

step = UInt8(16);
h = fit(Histogram, (vec(fixed), vec(moving)), (0x00:step:0xff, 0x00:step:0xff));
@benchmark append!(h, (vec(fixed), vec(moving)))
```

Baseline:
```
BenchmarkTools.Trial: 
  memory estimate:  192 bytes
  allocs estimate:  5
  --------------
  minimum time:     859.135 μs (0.00% GC)
  median time:      924.734 μs (0.00% GC)
  mean time:        959.305 μs (0.00% GC)
  maximum time:     1.570 ms (0.00% GC)
  --------------
  samples:          5208
  evals/sample:     1
```

This patch without my fix:
```
ERROR: InexactError: trunc(UInt8, 256)
Stacktrace:
 [1] throw_inexacterror(::Symbol, ::Type{UInt8}, ::Int64) at ./boot.jl:558
 [2] checked_trunc_uint at ./boot.jl:588 [inlined]
 [3] toUInt8 at ./boot.jl:650 [inlined]
 [4] UInt8 at ./boot.jl:710 [inlined]
 [5] convert at ./number.jl:7 [inlined]
 [6] getindex at ./range.jl:658 [inlined]
 [7] _edge_binindex at /home/salmon/Documents/StatsBase.jl/src/hist.jl:255 [inlined]
 [8] (::StatsBase.var"#131#132"{Histogram{Int64,2,Tuple{StepRange{UInt8,UInt8},StepRange{UInt8,UInt8}}}})(::StepRange{UInt8,UInt8}, ::UInt8) at /home/salmon/Documents/StatsBase.jl/src/hist.jl:227
 [9] map at ./tuple.jl:177 [inlined]
 [10] binindex at /home/salmon/Documents/StatsBase.jl/src/hist.jl:226 [inlined]
 [11] push! at /home/salmon/Documents/StatsBase.jl/src/hist.jl:311 [inlined]
 [12] append!(::Histogram{Int64,2,Tuple{StepRange{UInt8,UInt8},StepRange{UInt8,UInt8}}}, ::Tuple{Array{UInt8,1},Array{UInt8,1}}) at /home/salmon/Documents/StatsBase.jl/src/hist.jl:332
 [13] ##core#853() at /home/salmon/.julia/packages/BenchmarkTools/eCEpo/src/execution.jl:371
 [14] ##sample#854(::BenchmarkTools.Parameters) at /home/salmon/.julia/packages/BenchmarkTools/eCEpo/src/execution.jl:377
 [15] _run(::BenchmarkTools.Benchmark{Symbol("##benchmark#852")}, ::BenchmarkTools.Parameters; verbose::Bool, pad::String, kwargs::Base.Iterators.Pairs{Symbol,Integer,NTuple{4,Symbol},NamedTuple{(:samples, :evals, :gctrial, :gcsample),Tuple{Int64,Int64,Bool,Bool}}}) at /home/salmon/.julia/packages/BenchmarkTools/eCEpo/src/execution.jl:405
 [16] (::Base.var"#inner#2"{Base.Iterators.Pairs{Symbol,Integer,NTuple{5,Symbol},NamedTuple{(:verbose, :samples, :evals, :gctrial, :gcsample),Tuple{Bool,Int64,Int64,Bool,Bool}}},typeof(BenchmarkTools._run),Tuple{BenchmarkTools.Benchmark{Symbol("##benchmark#852")},BenchmarkTools.Parameters}})() at ./essentials.jl:713
 [17] #invokelatest#1 at ./essentials.jl:714 [inlined]
 [18] #run_result#37 at /home/salmon/.julia/packages/BenchmarkTools/eCEpo/src/execution.jl:32 [inlined]
 [19] run(::BenchmarkTools.Benchmark{Symbol("##benchmark#852")}, ::BenchmarkTools.Parameters; progressid::Nothing, nleaves::Float64, ndone::Float64, kwargs::Base.Iterators.Pairs{Symbol,Integer,NTuple{5,Symbol},NamedTuple{(:verbose, :samples, :evals, :gctrial, :gcsample),Tuple{Bool,Int64,Int64,Bool,Bool}}}) at /home/salmon/.julia/packages/BenchmarkTools/eCEpo/src/execution.jl:94
 [20] #warmup#45 at /home/salmon/.julia/packages/BenchmarkTools/eCEpo/src/execution.jl:141 [inlined]
 [21] warmup(::BenchmarkTools.Benchmark{Symbol("##benchmark#852")}) at /home/salmon/.julia/packages/BenchmarkTools/eCEpo/src/execution.jl:141
 [22] top-level scope at /home/salmon/.julia/packages/BenchmarkTools/eCEpo/src/execution.jl:287
```

This patch with my fix:
```
BenchmarkTools.Trial: 
  memory estimate:  192 bytes
  allocs estimate:  5
  --------------
  minimum time:     1.565 ms (0.00% GC)
  median time:      1.671 ms (0.00% GC)
  mean time:        1.679 ms (0.00% GC)
  maximum time:     2.385 ms (0.00% GC)
  --------------
  samples:          2977
  evals/sample:     1
```