@Moelf 2021-08-28
implemented in https://github.com/Moelf/FHist.jl